+++ src/style/style.ts
@@ -19,11 +19,7 @@ import {SourceCache} from '../source/source_cache';
 import {GeoJSONSource} from '../source/geojson_source';
 import {latest as styleSpec, derefLayers as deref, emptyStyle, diff as diffStyles, operations as diffOperations} from '@maplibre/maplibre-gl-style-spec';
 import {getGlobalWorkerPool} from '../util/global_worker_pool';
-import {
-    registerForPluginStateChange,
-    evented as rtlTextPluginEvented,
-    triggerPluginCompletionEvent
-} from '../source/rtl_text_plugin';
+import {rtlMainThreadPluginFactory} from '../source/rtl_text_plugin_main_thread';
 import {PauseablePlacement} from './pauseable_placement';
 import {ZoomHistory} from './zoom_history';
 import {CrossTileSymbolIndex} from '../symbol/cross_tile_symbol_index';
@@ -214,7 +210,6 @@ export class Style extends Evented {
     sourceCaches: {[_: string]: SourceCache};
     zoomHistory: ZoomHistory;
     _loaded: boolean;
-    _rtlTextPluginCallback: (a: any) => any;
     _changed: boolean;
     _updatedSources: {[_: string]: 'clear' | 'reload'};
     _updatedLayers: {[_: string]: true};
@@ -265,34 +260,7 @@ export class Style extends Evented {
         this._resetUpdates();
 
         this.dispatcher.broadcast('setReferrer', getReferrer());
-
-        const self = this;
-        this._rtlTextPluginCallback = registerForPluginStateChange((event) => {
-            const state = {
-                pluginStatus: event.pluginStatus,
-                pluginURL: event.pluginURL
-            };
-            self.dispatcher.broadcast('syncRTLPluginState', state)
-                .then((results) => {
-                    triggerPluginCompletionEvent(undefined);
-                    if (!results) {
-                        return;
-                    }
-                    const allComplete = results.every((elem) => elem);
-                    if (!allComplete) {
-                        return;
-                    }
-                    for (const id in self.sourceCaches) {
-                        const sourceType = self.sourceCaches[id].getSource().type;
-                        if (sourceType === 'vector' || sourceType === 'geojson') {
-                            // Non-vector sources don't have any symbols buckets to reload when the RTL text plugin loads
-                            // They also load more quickly, so they're more likely to have already displaying tiles
-                            // that would be unnecessarily booted by the plugin load event
-                            self.sourceCaches[id].reload(); // Should be a no-op if the plugin loads before any tiles load
-                        }
-                    }
-                }).catch((err: string) => triggerPluginCompletionEvent(err));
-        });
+        rtlMainThreadPluginFactory().on('pluginStateChange', this._rtlTextPluginStateChange);
 
         this.on('data', (event) => {
             if (event.dataType !== 'source' || event.sourceDataType !== 'metadata') {
@@ -317,6 +285,18 @@ export class Style extends Evented {
             }
         });
     }
+
+    _rtlTextPluginStateChange = () => {
+        for (const id in this.sourceCaches) {
+            const sourceType = this.sourceCaches[id].getSource().type;
+            if (sourceType === 'vector' || sourceType === 'geojson') {
+                // Non-vector sources don't have any symbols buckets to reload when the RTL text plugin loads
+                // They also load more quickly, so they're more likely to have already displaying tiles
+                // that would be unnecessarily booted by the plugin load event
+                this.sourceCaches[id].reload(); // Should be a no-op if the plugin loads before any tiles load
+            }
+        }
+    };
 
     loadURL(url: string, options: StyleSwapOptions & StyleSetterOptions = {}, previousStyle?: StyleSpecification) {
         this.fire(new Event('dataloading', {dataType: 'style'}));
@@ -1469,7 +1449,7 @@ export class Style extends Evented {
             this._spriteRequest.abort();
             this._spriteRequest = null;
         }
-        rtlTextPluginEvented.off('pluginStateChange', this._rtlTextPluginCallback);
+        rtlMainThreadPluginFactory().off('pluginStateChange', this._rtlTextPluginStateChange);
         for (const layerId in this._layers) {
             const layer: StyleLayer = this._layers[layerId];
             layer.setEventedParent(null);