+++ src/util/web_worker_transfer.ts
@@ -117,7 +117,8 @@ export function serialize(input: unknown, transferables?: Array<Transferable> | 
         input instanceof String ||
         input instanceof Date ||
         input instanceof RegExp ||
-        input instanceof Blob) {
+        input instanceof Blob ||
+        input instanceof Error) {
         return input;
     }
 
@@ -219,6 +220,7 @@ export function deserialize(input: Serialized): unknown {
         input instanceof Date ||
         input instanceof RegExp ||
         input instanceof Blob ||
+        input instanceof Error ||
         isArrayBuffer(input) ||
         isImageBitmap(input) ||
         ArrayBuffer.isView(input) ||