+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## main
 
 ### ✨ Features and improvements
+
 - ⚠️ Moved the `addSrouceType` to be a part of the global maplibregl object instead of being per map object ([#3420](https://github.com/maplibre/maplibre-gl-js/pull/3420))
+- ⚠️ Changed the `setRTLTextPlugin` function to return a promise instead of using callback ([#3418](https://github.com/maplibre/maplibre-gl-js/pull/3418)) this also changed how the RTL pluing code is handled internally by splitting the main thread and worker thread code.
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes